% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{AOO_summary}
\alias{AOO_summary}
\title{Summarise Area of Occurrence metrics}
\usage{
AOO_summary(AOO_grid)
}
\arguments{
\item{AOO_grid}{The AOO grid created by function \code{create_AOO_grid}}
}
\value{
A table with a summary of results.
}
\description{
Summarise Area of Occurrence metrics
}
\examples{
AOO_grid <- create_AOO_grid(glaciers_on_volcanos)
AOO_summary(AOO_grid)
}
