% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thresholds.R
\name{B_conditions}
\alias{B_conditions}
\title{Auxiliary for calculation of threshold for criterion B}
\usage{
B_conditions(
  spatial = FALSE,
  environment = FALSE,
  interactions = FALSE,
  threats = FALSE,
  locations = NA,
  ...
)
}
\arguments{
\item{spatial}{observed or inferred decline}

\item{environment}{observed or inferred decline}

\item{interactions}{observed or inferred decline}

\item{threats}{threatening process}

\item{locations}{number of locations}

\item{...}{additional parameters (ignored)}
}
\value{
A list with components named as the arguments.
}
\description{
Auxiliary for calculation of threshold for criterion B
}
