% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{create_AOO_grid}
\alias{create_AOO_grid}
\title{Create an Area of Occurrence grid}
\usage{
create_AOO_grid(pols, buffsize = 50000, cellsize = 10000, jitter = FALSE)
}
\arguments{
\item{pols}{A simple feature with Polygons (or Multipolygons).}

\item{buffsize}{Size of the buffer in meters.}

\item{cellsize}{Cell size in meter.}

\item{jitter}{Logical, should we add random displacement?}
}
\value{
A simple feature of a grid over the input polygons.
}
\description{
Create an Area of Occurrence grid
}
\examples{
AOO_grid <- create_AOO_grid(glaciers_on_volcanos)
print(AOO_grid)
}
