% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{create_AOO_simple_grid}
\alias{create_AOO_simple_grid}
\title{Create an Area of Occurrence grid for multiple ecosystems}
\usage{
create_AOO_simple_grid(pols, cellsize = 10000, names_from = NA)
}
\arguments{
\item{pols}{A simple feature with Points, Lines, Polygons or Multipolygons.}

\item{cellsize}{Cell size in meters.}

\item{names_from}{Which column includes the ecosystem names? if missing, will use a generic name}
}
\value{
A simple feature of a grid over the input polygons.
}
\description{
Create an Area of Occurrence grid for multiple ecosystems
}
