% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosystem_vector_map.R
\name{sf_as_ecomap}
\alias{sf_as_ecomap}
\title{Create ecomap object from an sf object}
\usage{
sf_as_ecomap(x, names_from = NA, area_from = NA, output_units = "km2")
}
\arguments{
\item{x}{\code{sf} object representing the distribution of one or more ecosystems}

\item{names_from}{Which column includes the ecosystem names? if missing, will use a generic name}

\item{area_from}{Which column includes the ecosystem extent (area)? if missing, will try to calculate it}

\item{output_units}{Set units for area output.}
}
\value{
ecosystem vector map
}
\description{
Create ecomap object from an sf object
}
\examples{
require(sf)
glaciers_on_volcanos <- tropical_glaciers |>
  dplyr::filter(ecosystem_name \%in\% "Volcanos de Peru y Chile") |>
  sf::st_transform(crs = 32719)
class(glaciers_on_volcanos)
glaciers_on_volcanos <- sf_as_ecomap(glaciers_on_volcanos)
class(glaciers_on_volcanos)
glaciers_on_volcanos
}
