% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{summary.AOO_grid}
\alias{summary.AOO_grid}
\title{Summarise Area of Occurrence metrics}
\usage{
\method{summary}{AOO_grid}(object, output_units = "km2", ...)
}
\arguments{
\item{object}{The AOO grid created by function \code{create_AOO_grid}}

\item{output_units}{set units for area output.}

\item{...}{further arguments passed to or from other methods, currently ignored.}
}
\value{
A table with a summary of results.
}
\description{
Summarise Area of Occurrence metrics
}
\examples{
AOO_grid <- create_AOO_grid(glaciers_on_volcanos)
summary(AOO_grid)
}
