% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EOO.R
\name{thresholds.EOO_convex_hull}
\alias{thresholds.EOO_convex_hull}
\title{Threshold for IUCN RLE criterion B1}
\usage{
\method{thresholds}{EOO_convex_hull}(x, ecosystem_name = NA, conditions = NULL, useNT = TRUE, ...)
}
\arguments{
\item{x}{The EOO convex hull created by function \code{create_EOO_chull}}

\item{ecosystem_name}{Which ecosystem to use? (if multiple ecosystems are present in the grid)}

\item{conditions}{list of conditions considered when applying the criterion}

\item{useNT}{logical, should we apply rules for the Near Threatened category? TRUE by default, if FALSE the category Least Concern will be used, but a note will be added to the output. See details.}

\item{...}{further arguments passed to B_conditions() if conditions is not provided}
}
\value{
Category of risk of collapse for subcriterion B1
}
\description{
Threshold for IUCN RLE criterion B1
}
