% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thresholds.R
\name{thresholds}
\alias{thresholds}
\title{Apply IUCN Red List of Ecosystem thresholds}
\usage{
thresholds(x, ...)
}
\arguments{
\item{x}{Object representing a spatial or functional metric}

\item{...}{further arguments passed to or from other methods}
}
\value{
A category
}
\description{
Apply IUCN Red List of Ecosystem thresholds
}
