% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EOO.R
\name{create_EOO_chull}
\alias{create_EOO_chull}
\title{Create an Extent of Occurrence convex hull}
\usage{
create_EOO_chull(pols)
}
\arguments{
\item{pols}{A simple feature with Polygons (or Multipolygons).}
}
\value{
A simple feature of a convex hull over the input polygons.
}
\description{
Create an Extent of Occurrence convex hull
}
