% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EOO.R
\name{print.EOO_convex_hull}
\alias{print.EOO_convex_hull}
\title{Print method for EOO convex hull}
\usage{
\method{print}{EOO_convex_hull}(x, ...)
}
\arguments{
\item{x}{the EOO convex hull}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for EOO convex hull
}
