% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{thresholds.AOO_grid}
\alias{thresholds.AOO_grid}
\title{Threshold for IUCN RLE criterion B2}
\usage{
\method{thresholds}{AOO_grid}(
  x,
  spatial = FALSE,
  environment = FALSE,
  interactions = FALSE,
  threat = FALSE,
  locations = NA,
  ...
)
}
\arguments{
\item{x}{The AOO grid created by function \code{create_AOO_grid}}

\item{spatial}{observed or inferred decline}

\item{environment}{observed or inferred decline}

\item{interactions}{observed or inferred decline}

\item{threat}{threatening process}

\item{locations}{number of locations}

\item{...}{further arguments passed to or from other methods}
}
\value{
B2 categories
}
\description{
Threshold for IUCN RLE criterion B2
}
\examples{
AOO_grid <- create_AOO_grid(glaciers_on_volcanos)
thresholds(AOO_grid)
}
