% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AOO.R
\name{thresholds.AOO_grid}
\alias{thresholds.AOO_grid}
\title{Threshold for IUCN RLE criterion B2}
\usage{
\method{thresholds}{AOO_grid}(x, rule = c("marginal", "small", "all"), conditions = NA, ...)
}
\arguments{
\item{x}{The AOO grid created by function \code{create_AOO_grid}}

\item{conditions}{list of conditions considered when applying the criterion}

\item{...}{further arguments passed to B_conditions() if conditions is not provided}
}
\value{
B2 categories
}
\description{
Threshold for IUCN RLE criterion B2
}
\examples{
AOO_grid <- create_AOO_grid(glaciers_on_volcanos)
thresholds(AOO_grid)
}
